+++ springapp/src/main/java/com/examly/springapp/service/review/ReviewService.java
@@ -1,0 +1,13 @@
+package com.examly.springapp.service.review;
+
+import java.util.List;
+import com.examly.springapp.models.ReviewModel;
+
+public interface ReviewService {
+
+    public ReviewModel addTask(ReviewModel mod);
+
+    public List<ReviewModel> getTask();
+
+    public boolean deleteTask(Long id);
+}
